Add recurse => true parameters to the directory resources for directories in Icinga's objects directory.
As covered in https://tickets.puppetlabs.com/browse/PUP-1327, Puppet's built-in Nagios types create .cfg files with root:root and the owner/group and 600 as the file mode. The recurse => true parameters on the directory resources makes Puppet set them back to what they should be for Icinga to be able to read them. This is likely just a temporary workaround until PUP-1327 is fixed.